aerc: add insert address capabilities to the editor
WIP. Use fzf to insert addresses into vim. Each address will go on a new
line with trailing commas. With the cursor on for example the empty To:
header and calling the function with multiple address should generate
the address in a way that they should work out of the box with aerc.