git: add aliases to better manage looking at remote diffs
Running set-head -a [remote] sets refs/remotes/[remote]/HEAD to whatever
is the default branch of the remote. This allows the `lm` to a) not fail
due to NOT having HEAD set (it's only set on a clone, not if you git
init and then git remote add) and will be independent of the default
branch being master/main/dev/etc.