aerc: remove the need for syncnotmuchtag.py
Originally the syncnotmuchtag.py was required to be able to move
messages from the unified notmuch view into other folders (e.g.
archive), but still keep in the appropriate account's maildir folder.
Utilize the new {{.Filename}} template, that incidentally allows you to
grab the appropriate maildir account's name (if the maildir folder on
the filesystem has the same name as the account name in aerc). This
allows moving with :mv [specificaccount]/folder. Unfortunately, this
does not work for gmail due to multiple messages.