imap: fix o365 disconnect
The current version of the config fill time-out as soon as the access
token is timed out and it will not be able to refresh, nor will it try
to fetch a new access token. Aerc has the ability to use refresh tokens
itself. Unfortunately, it's a mess, but we'll use it anyway.