ipython: add clipboard support
Putting the output of pandas DataFrames is a bit convoluted, if you need
to copy paste from the REPL to the file, especially if the DataFrame
does not fit into a terminal window. This sort of goes for any larger
text output. Add a clip function to ipython, that either reads the last
REPL output (the largest key from Out) or it's argument and puts it on
the clipboard. Currently, it only tries to be smart about pandas
DataFrames, where it call to_markdown as the representation.