email: move config to throttle
Move mbsync and notmuch logic from mailsync-high to aerc-hooks. Move the
mark eveything as read logic and msmstp-q retries to mailsync-medium.
Mailsync-high is now obsolete. The medium and low logic could also be
considered to be passed to throttle, basically just for the logging,
notifications and retries, but maybe a better solution would be
implementing systemd notifications.