throttle: protect against notmuch query folders
Ad-hoc query folders in notmuch can have very strange names and it's
currently impossible to figure out _what_ should be synced (except for
everything). Since this only affects flag-changed hook, so read/unread,
I don't think it's worth to sync everything in the unlikely chance this
is the way I read an email for the first time.