notes: alias to retore notes on reuse message
Sometimes when editing a commit it's easier to just git reset HEAD~,
than to amend it. Restoring the original message can be done by using
commit --reuse-message, but this still looses the notes. Change the ch
alias so it also restores any notes (note that if there are no notes,
this will print an ugly error message now, but hey, shouldn't be that
irritating).

Also figured out that you can alias an alias, so since I'm guessing I'll
be forgetting what ch is: add a more readable alias.