git: never forget a S-o-b again for git-send-email
In my experience S-o-b is only required for projects that also require
sending patches via email (by my experience I mean whenever I actually
needed to submit a patch). This is a safe way to not have a meaningless
S-o-b in e.g. work related projects that do not attach a meaning to
this.