cacyhos: lock screen on suspend
xss-lock executes screen lock on suspend. There is a slight delay
between opening the laptop and the lock kicking in unfortunately. I'm
guessing the suspend happens too fast to launch the lock _before_ the
suspend, so it is actually only launched after, but it is much better
than on Ubuntu.