keyboards: add mac layouts
I had some trouble with the Hungarian Mac keyboard layout, since it was
different than what I was used to on Linux/Windows. It also gave
a revelation, that I actually need two types. One that is the classing
Hungarian layout for typing (and writing Hungarian), and one where the
directly accessible accented characters are assigned characters more
often used in programing. Where this makes sense, I followed the English
layout, but since I didn't change the base Hungrian layout for this, it
did not make sense everywhere.