Revert "systemd: add linux homebrew to path"
This reverts commit c0fb9eeb7642d831186c6ba63f325ed84b16e55c.

Homebrew needs other things as well (hence the eval in shell). The
problem is that if homebrew is somewhere else (e.g. on Mac), the extra
PATH entry is not an issue, but the other variable seem to be. Of
course, probably any interactive session would overwrite this anyway,
but for now, let's remove it.