tasksync: add homebrew if it exists
On Ubuntu it's a pain to have certain packages up-to-date, hence
homebrew, but that needs to be eval-ed. Aur is a bit better than
homebrew since it has more packages and is seamless of course, not like
this eval, but on the downside, aur updates _everything_ which is
another kind of pain.

Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>